requirements.txt
flask>=3.0.0
psycopg2-binary